Transform test steps into task and steps.
