+++ lib/bin/roboter.js
@@ -28,6 +28,8 @@ const exit = function (ex) {
     return;
   }
 
+  console.log({ ex });
+
   if (ex.message) {
     buntstift.verbose(ex.message);
   }
