+++ lib/cli/commands/analyse.js
@@ -51,6 +51,10 @@ const analyse = {
   },
 
   async runTasks ({ directory }) {
+    if (!directory) {
+      throw new Error('Directory is missing.');
+    }
+
     await analyseCodeStep({ directory, ui });
   }
 };
