+++ lib/cli/commands/qa.js
@@ -4,7 +4,7 @@ const analyseTask = require('../tasks/analyse'),
       checkLicenseCompatibilityTask = require('../tasks/checkLicenseCompatibility'),
       dependenciesTask = require('../tasks/dependencies'),
       globalOptionDefinitions = require('../globalOptionDefinitions'),
-      testSteps = require('../../steps/test'),
+      testCodeTask = require('../tasks/testCode'),
       ui = require('../ui');
 
 const deps = {
@@ -33,23 +33,7 @@ const deps = {
     ui.printTaskHeader('quality assurance');
 
     await analyseTask({ directory, ui });
-
-    try {
-      await testSteps.testCode({ directory, ui });
-    } catch (ex) {
-      switch (ex.code) {
-        case 'ETESTSMISSING':
-          ui.warn('No tests found.');
-          break;
-        case 'ETESTSFAILED':
-          ui.printTaskFailure('Tests failed.');
-          throw ex;
-        default:
-          ui.printTaskFailure('Failed to run tests.');
-          throw ex;
-      }
-    }
-
+    await testCodeTask({ directory, ui });
     await dependenciesTask({ directory, ui });
     await checkLicenseCompatibilityTask({ directory, ui });
