+++ lib/cli/commands/release.js
@@ -9,7 +9,7 @@ const analyseTask = require('../tasks/analyse'),
       dependenciesTask = require('../tasks/dependencies'),
       files = require('../../files'),
       globalOptionDefinitions = require('../globalOptionDefinitions'),
-      testSteps = require('../../steps/test'),
+      testCodeTask = require('../tasks/testCode'),
       ui = require('../ui');
 
 const release = {
@@ -60,23 +60,7 @@ const release = {
     try {
       if (!force) {
         await analyseTask({ directory, ui });
-
-        try {
-          await testSteps.testCode({ directory, ui });
-        } catch (ex) {
-          switch (ex.code) {
-            case 'ETESTSMISSING':
-              ui.warn('No tests found.');
-              break;
-            case 'ETESTSFAILED':
-              ui.printTaskFailure('Tests failed.');
-              throw ex;
-            default:
-              ui.printTaskFailure('Failed to run tests.');
-              throw ex;
-          }
-        }
-
+        await testCodeTask({ directory, ui });
         await dependenciesTask({ directory, ui });
         await checkLicenseCompatibilityTask({ directory, ui });
       }
