+++ lib/cli/commands/test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const globalOptionDefinitions = require('../globalOptionDefinitions'),
-      testSteps = require('../../steps/test'),
+      testCodeTask = require('../tasks/testCode'),
       ui = require('../ui'),
       watchFilesAndExecute = require('../../watchFilesAndExecute');
 
@@ -52,31 +52,21 @@ const testCommand = {
     if (watch) {
       await watchFilesAndExecute({
         message: 'Running tests in watch mode...',
-        files: await testSteps.testCode.getFiles({ directory }),
-        execute: async () => await testSteps.testCode({ directory, type, printSuccess: true, ui }),
+        files: await testCodeTask.getFiles({ directory }),
+        execute: async () => await this.runTasks({ directory, type }),
         executeOnStart: true
       });
     }
 
-    try {
-      await testSteps.testCode({ directory, type, ui });
+    await this.runTasks({ directory, type });
+  },
 
-      ui.printTaskSuccess('Tests successful.');
-    } catch (ex) {
-      switch (ex.code) {
-        case 'ETESTSMISSING':
-          ui.warn('No tests found.');
-
-          return;
-        case 'ETESTSFAILED':
-          ui.printTaskFailure('Tests failed.');
-          break;
-        default:
-          ui.error('Failed to run tests.');
-      }
-
-      throw ex;
+  async runTasks ({ directory, type }) {
+    if (!directory) {
+      throw new Error('Directory is missing.');
     }
+
+    await testCodeTask({ directory, type, ui });
   }
 };
