+++ lib/cli/tasks/checkLicenseCompatibility.js
@@ -59,7 +59,6 @@ const checkLicenseCompatibility = async function ({ directory, ui }) {
 
       ui.table(table);
       ui.warn('There are incompatible dependencies.');
-
       throw new errors.LicenseIncompatible();
     }
