+++ lib/errors.js
@@ -19,8 +19,8 @@ const errors = defekt({
   OutdatedDependencies: {},
   PackageJsonMissing: {},
   PendingChanges: {},
+  StepExecutionFailed: {},
   TestsFailed: {},
-  TestsMissing: {},
   TypeScriptCompilationFailed: {},
   TypeScriptOutputConfigurationMissing: {},
   UnusedDependencies: {}
