+++ lib/runCommand.js
@@ -6,7 +6,7 @@ const processenv = require('processenv').default,
 const errors = require('./errors'),
       ui = require('./cli/ui');
 
-const exec = async function (command, options = {}) {
+const runCommand = async function (command, options = {}) {
   if (!command) {
     throw new Error('Command is missing.');
   }
@@ -42,4 +42,4 @@ const exec = async function (command, options = {}) {
   });
 };
 
-module.exports = exec;
+module.exports = runCommand;
