+++ lib/steps/test/runLifecycleStep.js
@@ -5,17 +5,15 @@ const path = require('path');
 const processenv = require('processenv').default,
       shell = require('shelljs');
 
-const files = require('../../files');
+const errors = require('../../errors'),
+      files = require('../../files');
 
-const runLifecycleStep = async function ({ directory, step, ui, environmentVariables = {}}) {
+const runLifecycleStep = async function ({ directory, step, environmentVariables = {}}) {
   if (!directory) {
     throw new Error('Directory is misssing.');
   }
   if (!step) {
     throw new Error('Step is misssing.');
-  }
-  if (!ui) {
-    throw new Error('UI is misssing.');
   }
 
   const jsStepFile = path.join(directory, `${step}.js`);
@@ -26,8 +24,6 @@ const runLifecycleStep = async function ({ directory, step, ui, environmentVaria
   if (!jsStepExists && !tsStepExists) {
     return;
   }
-
-  ui.printTaskHeader(`${step} step`);
 
   const mergedEnvironmentVariables = {
     ...processenv(),
@@ -41,7 +37,7 @@ const runLifecycleStep = async function ({ directory, step, ui, environmentVaria
     });
 
     if (jsStep.code !== 0) {
-      throw new Error(`${step} step execution failed.`);
+      throw new errors.StepExecutionFailed(`${step} step execution failed.`);
     }
   } else if (tsStepExists) {
     const tsPre = shell.exec(`ts-node -r tsconfig-paths/register ${tsStepFile}`, {
@@ -50,7 +46,7 @@ const runLifecycleStep = async function ({ directory, step, ui, environmentVaria
     });
 
     if (tsPre.code !== 0) {
-      throw new Error(`${step} step execution failed.`);
+      throw new errors.StepExecutionFailed(`${step} step execution failed.`);
     }
   }
 };
